+++ clisops/core/subset.py
@@ -580,8 +580,120 @@ def create_mask(
     return mask
 
 
-def grid_polygon(ds):
-    """Return a polygon tracing the grid boundary.
+def _rectilinear_grid_exterior_polygon(ds):
+    """Return a polygon tracing a rectilinear grid's exterior.
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+      CF-compliant input dataset.
+
+    Returns
+    -------
+    shapely.geometry.Polygon
+      Grid cell boundary.
+    """
+
+    # Add bounds if not present
+    if 'longitude' not in ds.cf.bounds:
+        ds = ds.cf.add_bounds("longitude")
+    if 'latitude' not in ds.cf.bounds:
+        ds = ds.cf.add_bounds("latitude")
+
+    x = ds.cf.get_bounds("longitude")  # lon_bnds
+    y = ds.cf.get_bounds("latitude")  # lat_bnds
+
+    # Take the grid corner coordinates
+    xmin = x[0, 0]
+    xmax = x[-1, -1]
+    ymin = y[0, 0]
+    ymax = y[-1, -1]
+
+    pts = [(xmin, ymin), (xmin, ymax), (xmax, ymax), (xmax, ymin)]
+    return Polygon(pts)
+
+
+def _curvilinear_grid_exterior_polygon(ds, mode="bbox"):
+    """Return a polygon tracing a curvilinear grid's exterior.
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+      CF-compliant input dataset.
+    mode : {bbox, cell_union}
+      Calculation mode. `bbox` takes the min and max longitude and latitude bounds and rounds them to 0.1 degree.
+      `cell_union` merges all grid cell polygons and finds the exterior. Also rounds and simplifies the coordinates
+      to smooth projection errors.
+
+    Returns
+    -------
+    shapely.geometry.Polygon
+      Grid cell boundary.
+    """
+    from shapely.ops import unary_union
+    import math
+
+    def round_up(x, decimal=1):
+        f = 10 ** decimal
+        return math.ceil(x * f) / f
+
+    def round_down(x, decimal=1):
+        f = 10 ** decimal
+        return math.floor(x * f) / f
+
+    if mode == "bbox":
+        try:
+            # cf-convention
+            x = ds.cf.get_bounds("longitude")  # lon_bnds
+            y = ds.cf.get_bounds("latitude")  # lat_bnds
+        except KeyError:
+            # xesmf convention
+            x = ds.lon_b
+            y = ds.lat_b
+
+        xmin = round_down(x.min())
+        xmax = round_up(x.max())
+        ymin = round_down(y.min())
+        ymax = round_up(y.max())
+
+        pts = [(xmin, ymin), (xmin, ymax), (xmax, ymax), (xmax, ymin)]
+
+    elif mode == "cell_union":
+        # x and y should be vertices.
+        # There is no guarantee that the sides of the array storing the curvilinear grids corresponds to the exterior of
+        # the lon/lat grid.
+        # For example, in a polar stereographic projection, the pole would be at the center of the native grid.
+        # So we need to create individual polygons for each grid cell, take the union and get the exterior. Even then,
+        # for some grids, projection distortions might introduce errors.
+        # Consider this code experimental.
+
+        # If the following fails, it's probably because the axis attribute is not set for the coordinates.
+        xax = ds.cf.axes["X"][0]
+        yax = ds.cf.axes["Y"][0]
+
+        # Stack i and j
+        sds = ds.stack(zkz_=(xax, yax))
+
+        x = sds.cf.get_bounds("longitude")  # lon_bnds
+        y = sds.cf.get_bounds("latitude")  # lat_bnds
+
+        # Grid cell polygons
+        polys = [Polygon(zip(lx, ly)) for lx, ly in zip(x.data.T, y.data.T)]
+
+        # Exterior of all these polygons
+        pts = unary_union(polys).simplify(.1).buffer(.1).exterior
+        x, y = np.around(pts.xy, 1)
+        y = np.clip(y, -90, 90)
+        pts = zip(x, y)
+
+    return Polygon(pts)
+
+
+def grid_exterior_polygon(ds):
+    """Return a polygon tracing the grid's exterior.
+
+    This function is only accurate for a geographic lat/lon projection.
+    For projected grids, it's a rough approximation.
 
     Parameters
     ----------
@@ -599,42 +711,11 @@ def grid_polygon(ds):
     you need this.
     """
     from shapely.geometry import Polygon
-    from itertools import chain
 
     if is_rectilinear(ds):
-        if getattr(ds.cf.bounds, "X", None) not in ds:
-            lon_name = ds.cf["longitude"].name
-            lat_name = ds.cf["latitude"].name
-            ds = ds.cf.add_bounds([lon_name, lat_name])
+        return _rectilinear_grid_exterior_polygon(ds)
 
-        x = ds.cf.bounds["X"][0]  # lon_bnds
-        y = ds.cf.bounds["Y"][0]  # lat_bnds
-        xmin = ds[x][0, 0]
-        xmax = ds[x][-1, -1]
-        ymin = ds[y][0, 0]
-        ymax = ds[y][-1, -1]
-
-        pts = [(xmin, ymin), (xmin, ymax), (xmax, ymax), (xmax, ymin)]
-
-    else:
-        # This is a partial implementation, in that it doesn't use the grid boundaries, but the centroids.
-        xax = ds.cf.axes["X"][0]
-        yax = ds.cf.axes["Y"][0]
-
-        lon = ds.cf["longitude"]
-        lat = ds.cf["latitude"]
-        coords = xarray.Dataset({"lon": lon, "lat": lat})
-
-        sides = [
-            coords.isel({xax: 0}),
-            coords.isel({yax: -1}),
-            coords.isel({xax: -1}),
-            coords.isel({yax: 0}),
-        ]
-
-        pts = chain(*[zip(side.lon.data, side.lat.data) for side in sides])
-
-    return Polygon(pts)
+    return _curvilinear_grid_exterior_polygon(ds, mode="bbox")
 
 
 def is_rectilinear(ds):
@@ -677,19 +758,18 @@ def shape_bbox_indexer(ds, poly):
     if rectilinear:
         envelope = geom.minimum_rotated_rectangle
     else:
-        # Convex hull - for curvilinear grids, this seems safer than the rotated_rectangle
+        # For curvilinear grids, the convex hull seems safer than the rotated_rectangle.
         envelope = geom.convex_hull
 
-    # If polygon sits on the grid boundary, we need to roll the grid's coordinates.
-    if not grid_polygon(ds).contains(envelope):
-        # This feature is not supported for now. Please submit a PR if this is an itch you want to scratch.
+    # If polygon sits on the grid boundary, we need to roll the grid's coordinates and this is not supported.
+    if not grid_exterior_polygon(ds).contains(envelope):
         return {}
 
-    # Create index from edge vertices
+    # Create index from edge vertices (last item is just a copy of the first to close the polygon)
     elon, elat = map(np.array, zip(*envelope.boundary.coords[:-1]))
 
-    # Create envelope coordinates (last item is just a copy of the first to close the polygon)
-    ind = {ds.cf["longitude"].name: elon[:-1], ds.cf["latitude"].name: elat[:-1]}
+    # Create envelope coordinates
+    ind = {ds.cf["longitude"].name: elon, ds.cf["latitude"].name: elat}
 
     # Find indices nearest the rectangle' corners
     # Note that the nearest indices might be inside the shape, so we'll need to add a *halo* around those indices.
@@ -700,6 +780,7 @@ def shape_bbox_indexer(ds, poly):
 
     else:
         # For curvilinear grids, finding the closest points require a bit more work.
+        # Note that this code is not exercised for now.
         from scipy.spatial import cKDTree
 
         # These are going to be 2D grids.
@@ -715,7 +796,7 @@ def shape_bbox_indexer(ds, poly):
 
     # Create slices, adding a halo around selection to account for `nearest` grid cell center approximation.
     out = {}
-    halo = 1
+    halo = 2
     for (k, v) in native_ind.items():
         vmin = np.clip(v.min() - halo, 0, ds[k].size)
         vmax = np.clip(v.max() + halo + 1, 0, ds[k].size)